+++ adlfs/spec.py
@@ -1285,7 +1285,32 @@ class AzureBlobFileSystem(AsyncFileSystem):
         for file in file_paths:
             self.invalidate_cache(self._parent(file))
 
-    rm_files = sync_wrapper(_rm_files)
+    sync_wrapper(_rm_files)
+
+    async def _rm_file(
+        self, path: typing.Union[str, typing.List[str]], delimiter: str = "/", **kwargs
+    ):
+        """Delete a file.
+
+        Parameters
+        ----------
+        path: str
+            File to delete.
+        """
+        container_name, p, _ = self.split_path(path, delimiter=delimiter)
+        try:
+            if p != "":
+                await self._rm_files(container_name, [p.rstrip(delimiter)])
+            else:
+                await self._rmdir(container_name)
+        except ResourceNotFoundError:
+            pass
+        except FileNotFoundError:
+            pass
+        except Exception as e:
+            raise RuntimeError("Failed to remove %s for %s", path, e) from e
+
+    rm_file = sync_wrapper(_rm_file)
 
     async def _separate_directory_markers_for_non_empty_directories(
         self, file_paths: typing.Iterable[str]